app/serializers.py
from rest_framework import serializers
from app.models import *
from django.contrib.auth.models import User

class BlogSerializer(serializers.ModelSerializer):
    class Meta:
        model = Blog
        fields = ['id', 'image', 'created', 'title', 'content', 'author']
        read_only_fields = ['author', 'created']

# Create your tests here.
class UserSerializer(serializers.ModelSerializer):
    class Meta(object):
        model = User
        fields = ['id', 'username', 'password', 'email']
